% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brocolors.R
\name{brocolors}
\alias{brocolors}
\title{Vectors of colors for figures}
\usage{
brocolors(
  set = c("general", "bg", "bgpng", "CC", "f2", "sex", "main", "crayons")
)
}
\arguments{
\item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{

}
\seealso{
\code{\link{plot_crayons}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{utilities}
